googlesheet/requirements_sheets.txt
# Dependencias específicas para Google Sheets
pandas==2.2.2
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
gspread==5.12.0
gspread-dataframe==3.3.1
python-dotenv==1.0.1
